Add restore command
The LPC55 relies on having a valid image in ROM in order to attach via
SWD. Flash programming via SWD can fail if the image is in a lock up
state. Add a restore command to erase the flash and put a minimal set
of bytes in flash to allow attaching and flashing via SWD to work.