Allow setting the various DICE fields in the CMPA.
Names of variables / parameters have been kept as similar to those from
the spec as possible. These names are pretty awkward but we preserve
them as best we can for consistency.

The `sign_image` function has been split in two: One function to sign
the image and another to generate the CMPA.