Upgrade serialport-rs from 3.3.0 to 4.2.0.
This upgrade was mostly painless except for the new SerialPortBuilder.
Where the former interface accepted a path to the port as an OsStr the
new builder will only accept a parameter that implements
Into<std::borrow::Cow<'a, str>. This prevents us from using the PathBuf
where we store the port w/o doing a lossy conversion and handling a
possible error. As a solution this PR changes how we represent the port.
We now store it as a String, relying on clap to report any error
converting the input to this type.