Move off pre-release rsa crate that broke things
lpc55_sign was using the rsa 0.9.0-pre.0 crate, which (when people use
it as a library) would select the 0.9.0 released version. Unfortunately
the released version contains API incompatibilities with pre.0, breaking
builds. (This is one of the reasons why depending on prerelease crates
is a bad idea.)

This switches to the released 0.9.0 and fixes the incompatibilities that
resulted, which came down to a single trait being moved into a module.

I've bumped the crate version number so that people can depend on a
known-working version, and because recent commits have changed some
signing API incompatibly.