Change the default ISP setting in CMPA to UART.
The previous choice of "Auto" authorizes the ROM to auto-detect
connected interfaces. Since we know there's at least one exploitable bug
in the USB bootloader, we really don't want the USB interface to be
activated.

One could imagine making this a flag, but, given that we're hardcoding
our preferred clock frequency in the line above, I went for expedience.