lpc55_isp: don't require params to be Vecs
This loosens the parameter requirements to permit slices or arrays. If
the operation actually requires a Vec, it'll either create one
internally, or the caller can donate one. If the operation does _not_
require a Vec (like the data operations that chunk output before
sending), callers no longer have to allocate one.

I've converted callers in this repo to use the simplest types that
seemed reasonable in each case, which eliminates some extra allocation
in a very few cases.